Clean out prism tryouts and fix lint
