+++ composer.json
@@ -37,7 +37,6 @@
         "symfony/event-dispatcher": "^5.1",
         "symfony/filesystem": "^5.1",
         "symfony/finder": "^5.1",
-        "symfony/http-client": "^5.1",
         "symfony/http-foundation": "^5.1",
         "symfony/http-kernel": "^5.1",
         "symfony/process": "^5.1",
