+++ src/Highlighter/Prism.php
@@ -9,6 +9,7 @@
 namespace Content\Highlighter;
 
 use Content\Behaviour\HighlighterInterface;
+use Symfony\Component\Process\InputStream;
 use Symfony\Component\Process\Process;
 
 /**
@@ -17,10 +18,34 @@ use Symfony\Component\Process\Process;
 class Prism implements HighlighterInterface
 {
     private string $executable;
+    private ?Process $server = null;
+    private ?InputStream $input = null;
 
     public function __construct(string $executable = __DIR__ . '/../Resources/node/prism.js')
     {
         $this->executable = $executable;
+    }
+
+    public function start(): void
+    {
+        if (!$this->server) {
+            $this->input = new InputStream();
+            $this->server = new Process(['node', $this->executable], null, null, $this->input);
+        }
+
+        if (!$this->server->isRunning()) {
+            $this->server->start();
+        }
+    }
+
+    public function stop(): void
+    {
+        if (!$this->server || !$this->server->isRunning()) {
+            return;
+        }
+
+        $this->server->stop();
+        $this->input->close();
     }
 
     /**
@@ -28,14 +53,16 @@ class Prism implements HighlighterInterface
      */
     public function highlight(string $value, string $language): string
     {
-        $process = new Process(['node', $this->executable, $language, $value]);
+        $this->start();
 
-        $process->run();
+        $this->input->write(
+            json_encode(['language' => $language, 'value' => $value]) . PHP_EOL
+        );
 
-        if (!$process->isSuccessful()) {
-            throw new \RuntimeException($process->getErrorOutput());
-        }
+        $this->server->waitUntil(function ($type, $output) {
+            return $type === Process::ERR && $output === 'DONE';
+        });
 
-        return trim($process->getOutput());
+        return $this->server->getIncrementalOutput();
     }
 }
